md component builds and passes tests

+++ hoomd/md/NeighborListGPUStencil.h
@@ -75,7 +75,7 @@ class PYBIND11_EXPORT NeighborListGPUStencil : public NeighborListGPU
 
     private:
         std::unique_ptr<Autotuner> m_tuner;   //!< Autotuner for block size and threads per particle
-        unsigned int m_last_tuned_timestep;     //!< Last tuning timestep
+        uint64_t m_last_tuned_timestep;       //!< Last tuning timestep
 
         std::shared_ptr<CellList> m_cl;   //!< The cell list
         std::shared_ptr<CellListStencil> m_cls;   //!< The cell list stencil
